% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logit_result.R
\name{get_logit_result}
\alias{get_logit_result}
\title{Title Get Logit_model's Result}
\usage{
get_logit_result(logit_model, model_name)
}
\arguments{
\item{logit_model}{logit_model}

\item{model_name}{model's name}
}
\value{
a data.frame of logit_model's result
}
\description{
Title Get Logit_model's Result
}
\examples{
logit_model <- glm(y ~ x, data, family=familytype(link = 'binomial'))
logit_result <- get_logit_result(logit_model, 'model_1')


}
