% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_off.R
\name{turn_off}
\alias{turn_off}
\title{Title Turn Off EXE}
\usage{
turn_off(exe_name)
}
\arguments{
\item{exe_name}{}
}
\value{
nothing
}
\description{
Title Turn Off EXE
}
\examples{
turn_off(ppt)
turn_off(chrome)
}
